+++ kinto/views/records.py
@@ -1,10 +1,8 @@
 from kinto.core import resource, utils
-from kinto.core.errors import raise_invalid
 from pyramid.security import Authenticated
-from pyramid.settings import asbool
 
 from kinto.views import object_exists_or_404
-from kinto.schema_validation import validate_schema, ValidationError
+from kinto.schema_validation import validate_from_parent_schema_or_400
 
 
 _parent_path = '/buckets/{{bucket_id}}/collections/{{collection_id}}'
@@ -22,24 +20,15 @@ class Record(resource.ShareableResource):
         # Check if already fetched before (in batch).
         collections = request.bound_data.setdefault('collections', {})
         collection_uri = self.get_parent_id(request)
-        bucket_uri = utils.instance_uri(request, 'bucket', id=self.bucket_id)
         if collection_uri not in collections:
             # Unknown yet, fetch from storage.
+            bucket_uri = utils.instance_uri(request, 'bucket', id=self.bucket_id)
             collection = object_exists_or_404(request,
                                               collection_id='collection',
                                               parent_id=bucket_uri,
                                               object_id=self.collection_id)
             collections[collection_uri] = collection
         self._collection = collections[collection_uri]
-
-        buckets = request.bound_data.setdefault('buckets', {})
-        if bucket_uri not in buckets:
-            bucket = object_exists_or_404(request,
-                                          collection_id='bucket',
-                                          parent_id='',
-                                          object_id=self.bucket_id)
-            buckets[bucket_uri] = bucket
-        self._bucket = buckets[bucket_uri]
 
         super().__init__(request, **kwargs)
 
@@ -54,36 +43,17 @@ class Record(resource.ShareableResource):
         """Validate records against collection schema, if any."""
         new = super().process_record(new, old)
 
-        schemas = []
-        if 'schema' in self._collection:
-            schema_timestamp = self._collection[self.model.modified_field]
-            schemas.append(self._collection['schema'])
-        if 'record:schema' in self._bucket:
-            schema_timestamp = max(self._bucket[self.model.modified_field],
-                                   self._collection[self.model.modified_field])
-            schemas.append(self._bucket['record:schema'])
-
-        settings = self.request.registry.settings
-        schema_validation = 'experimental_collection_schema_validation'
-        if len(schemas) == 0 or not asbool(settings.get(schema_validation)):
-            return new
-
-        # Assign the schema version to the record.
-        new[self.schema_field] = schema_timestamp
-
         # Remove internal and auto-assigned fields from schemas and record.
         internal_fields = (self.model.id_field,
                            self.model.modified_field,
                            self.schema_field,
                            self.model.permissions_field)
-        data = {f: v for f, v in new.items() if f not in internal_fields}
-
-        for schema in schemas:
-            # Validate or fail with 400.
-            try:
-                validate_schema(data, schema, ignore_fields=internal_fields)
-            except ValidationError as e:
-                raise_invalid(self.request, name=e.field, description=e.message)
+        validate_from_parent_schema_or_400(new, resource_name="record", request=self.request,
+                                           ignore_fields=internal_fields)
+        # Assign the schema version to the record.
+        if 'schema' in self._collection:
+            schema_timestamp = self._collection[self.model.modified_field]
+            new[self.schema_field] = schema_timestamp
 
         return new